refactor: moved pipeline inside internal/
