+++ internal/http/service.go
@@ -30,7 +30,7 @@ import (
 	command "github.com/tarungka/wire/internal/command/proto"
 	"github.com/tarungka/wire/internal/logger"
 	"github.com/tarungka/wire/internal/new/store"
-	"github.com/tarungka/wire/pipeline"
+	"github.com/tarungka/wire/internal/pipeline"
 	"github.com/tarungka/wire/sinks"
 	"github.com/tarungka/wire/sources"
 )
