+++ internal/pipeline/worker.go
@@ -1,0 +1,9 @@
+package pipeline
+
+import "sync"
+
+
+type Worker struct {
+	pool sync.Pool
+	mu sync.Mutex
+}
